save 2 14/11

+++ Circulo de seguridad/Dtos/CrearEvento.cs
@@ -10,8 +10,8 @@ namespace Circulo_de_seguridad.Dtos
     {
         [Required]
         public string Identificador { get; set; }
-        public int CoordenadaX { get; set; }
-        public int CoordenadaY { get; set; }
+        public float CoordenadaX { get; set; }
+        public float CoordenadaY { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public string FechaFinalizacion { get; set; }
