+++ Circulo de seguridad/Entidades/Evento.cs
@@ -8,8 +8,8 @@ namespace Circulo_de_seguridad.Entidades
     public class Evento
     {
         public int Id { get; set; }
-        public int CoordenadaX { get; set; }
-        public int CoordenadaY { get; set; }
+        public float CoordenadaX { get; set; }
+        public float CoordenadaY { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public DateTime FechaCreacion { get; set; }
